+++ config/enquire.php
@@ -2,13 +2,19 @@
 
 return [
     'fields' => [
-        'text'        => 'enquire::partials.text',
-        'textarea'    => 'enquire::partials.textarea',
-        'email'       => 'enquire::partials.email',
-        'phone'       => 'enquire::partials.phone',
-        'file_upload' => 'enquire::partials.file',
-        'checkbox'    => 'enquire::partials.checkbox',
-        'dropdown'    => 'enquire::partials.dropdown',
+        'checkbox'      => 'enquire::partials.checkbox',
+        'dropdown'      => 'enquire::partials.dropdown',
+        'email'         => 'enquire::partials.email',
+        'file_upload'   => 'enquire::partials.file',
+        'phone'         => 'enquire::partials.phone',
+        'radio_buttons' => 'enquire::partials.radios',
+        'text'          => 'enquire::partials.text',
+        'textarea'      => 'enquire::partials.textarea',
+    ],
+
+    'fields_with_options' => [
+        'dropdown',
+        'radio_buttons',
     ],
 
     'from_address' => '',