+++ src/Forms/Fields/FieldDecorator.php
@@ -13,7 +13,6 @@ use Bozboz\Admin\Reports\Filters\HiddenFilter;
 use Bozboz\Admin\Reports\Filters\RelationFilter;
 use Bozboz\Enquire\Forms\Form;
 use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Support\Facades\Config;
 
 class FieldDecorator extends ModelAdminDecorator
 {
@@ -32,10 +31,9 @@ class FieldDecorator extends ModelAdminDecorator
 
 	public function getColumns($instance)
 	{
-		$type = studly_case(array_search($instance->input_type, Config::get('enquire.fields')));
 		return [
 			'Name' => $instance->label,
-			'Type' => $type,
+			'Type' => $instance->type_label,
 			'Required' => $instance->required ? '<i class="fa fa-check"></i>' : '',
 		];
 	}
@@ -57,6 +55,8 @@ class FieldDecorator extends ModelAdminDecorator
 	public function getFields($instance)
 	{
 		return [
+			new HiddenField(['name' => 'input_type']),
+			new TextField(['name' => 'type_label', 'disabled' => 'disabled']),
 			new TextField(['name' => 'label', 'label' => 'Name']),
 			new TextField(['name' => 'placeholder']),
 			new CheckboxField(['name' => 'required']),
@@ -65,11 +65,11 @@ class FieldDecorator extends ModelAdminDecorator
 				'data-tags' => 'true',
 				'help_text' => 'Enter validation rules for the field value',
 			]),
-			($instance->input_type == 'enquire::partials.dropdown' ? new TextareaField('options', [
-				'help_text' => 'Enter options for the dropdown with a new line between each one'
-			]) : null),
+			($instance->hasOptions() || $instance->input_type == 'enquire::partials.dropdown'
+				? new TextareaField('options', [
+					'help_text' => 'Enter options a new line between each one'
+				]) : null),
 			new HiddenField(['name' => 'form_id']),
-			new HiddenField(['name' => 'input_type']),
 		];
 	}
 }